Enhance product and sale management by adding SaleDetail model, updating product price type, and implementing JWT authentication for user login

+++ app/models/model_product.py
@@ -1,5 +1,5 @@
 from app.database.database import Base
-from sqlalchemy import Column, Integer, String, Date, ForeignKey
+from sqlalchemy import Column, Integer, String, Date, ForeignKey, Numeric
 
 
 class Product(Base):
@@ -8,7 +8,7 @@ class Product(Base):
     id_product = Column(Integer, primary_key=True, index=True)
     name = Column(String(100), nullable=False)
     description = Column(String(255), nullable=True)
-    price = Column(Integer, nullable=False)
+    price = Column(Numeric(10, 2), nullable=False)
     stock = Column(Integer, nullable=False)
     barcode = Column(String(100), nullable=True)
     created_at = Column(Date, nullable=False)
